\name{CleanUpdevs}
\alias{CleanUpdevs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ to come }
\description{ to come }
\usage{
CleanUpdevs()
}
%- maybe also 'usage' for other objects documented here.
\details{ }
\value{  }
\references{   }
\author{  Markus Loecher, Sense Networks <markus@sensenetworks.com>  }
\note{ }
\seealso{  }
\examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
