\name{HTMLhref}
\alias{HTMLhref}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ adds an href item to the current HTML page }
\description{adds an href item to the current HTML page }
\usage{
HTMLhref(href, txt, file = get(".HTML.file"), append = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{href}{ HTML reference/URL}
  \item{txt}{ text to display}
  \item{file}{ file to write to }
  \item{append}{ append to file (default TRUE) }
}
\details{ }
\value{  }
\references{   }
\author{  Markus Loecher, Sense Networks <markus@sensenetworks.com>  }
\note{ }
\seealso{  }
\examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
