\name{HTMLUtils-package}
\alias{HTMLUtils-package}
\docType{package}
\title{Facilitates automated HTML report creation}
\description{Facilitates automated HTML report creation, in particular framed HTML pages and dynamically sortable tables.}
\details{
\tabular{ll}{Package: \tab HTMLUtils\cr
Type: \tab Package\cr
Title: \tab Facilitates automated HTML report creation\cr
Version: \tab 0.1.5\cr
Date: \tab 2012-04-29\cr
Depends: \tab R2HTML\cr
Suggests: \tab \cr
Author: \tab "Markus Loecher, Berlin School of Economics and Law (BSEL)" <markus.loecher@gmail.com>\cr
Maintainer: \tab "Markus Loecher, Berlin School of Economics and Law (BSEL)" <markus.loecher@gmail.com>\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr}
}
\author{"Markus Loecher, Berlin School of Economics and Law (BSEL)" <markus.loecher@gmail.com>}

\keyword{ package }


