\name{InstallJSC}
\alias{InstallJSC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ installs the JS components  }
\description{prompts the user to install the JS components to the relevant directory, which enables dynamically sortable tables. }
\usage{
InstallJSC(JSCPATH)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{JSCPATH}{ path to install the jsc directory to. Recommended is the base public_html directory. }
}
\details{ }
\value{  }
\references{   }
\author{  Markus Loecher, Sense Networks <markus@sensenetworks.com>  }
\note{ }
\seealso{  }
\examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
