\name{makePathName}
\alias{makePathName}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ to come }
\description{ to come }
\usage{
makePathName(path, MakePath = TRUE, verbose = 0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{path}{ }
  \item{MakePath}{ }
  \item{verbose}{ }
}
\details{ }
\value{  }
\references{   }
\author{  Markus Loecher, Sense Networks <markus@sensenetworks.com>  }
\note{ }
\seealso{  }
\examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
