\name{MyReportBegin}
\alias{MyReportBegin}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{  gracefully initializes the HTML page }
\description{ gracefully initializes the HTML page }
\usage{
MyReportBegin(file = "report.html", title = "My Report Title", header = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{ filename }
  \item{title}{ title for HTML page}
  \item{header}{ header yes/no}
}
\details{ }
\value{  }
\references{   }
\author{  Markus Loecher, Sense Networks <markus@sensenetworks.com>  }
\note{ }
\seealso{  }
\examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
