\name{MyReportEnd}
\alias{MyReportEnd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ gracefully finalizes the HTML page }
\description{ gracefully finalizes the HTML page }
\usage{
MyReportEnd(file = "report.html")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{ }
}
\details{ }
\value{  }
\references{   }
\author{  Markus Loecher, Sense Networks <markus@sensenetworks.com>  }
\note{ }
\seealso{  }
\examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
