\name{CairoWorks}
\alias{CairoWorks}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ to come }
\description{ to come }
\usage{
CairoWorks(testfile = paste(tempdir(),"/tmp.png",sep=""), verbose = 0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{testfile}{ }
  \item{verbose}{ }
}
\details{ }
\value{  }
\references{   }
\author{  Markus Loecher, Sense Networks <markus@sensenetworks.com>  }
\note{ }
\seealso{  }
\examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
