\name{HTMLhref}
\alias{HTMLhref}
\title{adds an href item to the current HTML page}
\description{adds an href item to the current HTML page}
\usage{HTMLhref(href, txt, file = get(".HTML.file"), append = TRUE)}
\arguments{
  \item{href}{HTML reference/URL}
  \item{txt}{text to display}
  \item{file}{file to write to}
  \item{append}{append to file (default TRUE)}
}



\author{"Markus Loecher, Berlin School of Economics and Law (BSEL)"  <markus.loecher@gmail.com>}





