\name{HTMLUtils-package}
\alias{HTMLUtils-package}
\alias{HTMLUtils}
\docType{package}
\title{Facilitates Automated HTML Report Creation}
\description{Facilitates automated HTML report creation, in particular
framed HTML pages and dynamically sortable tables.}
\details{
\tabular{ll}{Package: \tab HTMLUtils\cr
Type: \tab Package\cr
Title: \tab Facilitates Automated HTML Report Creation\cr
Version: \tab 0.1.7\cr
Date: \tab 2015-01-17\cr
Depends: \tab R2HTML\cr
Suggests: \tab \cr
Author: \tab "Markus Loecher, Berlin School of Economics and Law (BSEL)"  <markus.loecher@gmail.com>\cr
Maintainer: \tab "Markus Loecher, Berlin School of Economics and Law (BSEL)" <markus.loecher@gmail.com>\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
Packaged: \tab 2012-05-17 21:56:35 UTC; mloecher\cr
Repository: \tab CRAN\cr
Date/Publication: \tab 2012-05-18 05:59:13\cr}
}
\author{"Markus Loecher, Berlin School of Economics and Law (BSEL)"  <markus.loecher@gmail.com>}

\keyword{ package }


