\name{HTMLUtils-package}
\alias{HTMLUtils-package}
\alias{HTMLUtils}
\docType{package}
\title{
Facilitates automated HTML report creation
}
\description{
Facilitates automated HTML report creation, in particular framed HTML pages and dynamically sortable tables.
}
\details{
\tabular{ll}{
Package: \tab HTMLUtils\cr
Type: \tab Package\cr
Version: \tab 0.1.4\cr
Date: \tab 2010-07-29\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{
Markus Loecher, Sense Networks <markus@sensenetworks.com> 
}
\references{

}
\seealso{
}
\examples{
}
