\name{InstallJSC}
\alias{InstallJSC}
\title{installs the JS components}
\description{prompts the user to install the JS components to the relevant directory, which enables dynamically sortable tables.}
\usage{InstallJSC(JSCPATH)}
\arguments{
  \item{JSCPATH}{path to install the jsc directory to. Recommended is the base public html directory.}
}



\author{"Markus Loecher, Berlin School of Economics and Law (BSEL)" <markus.loecher@gmail.com>}





