\name{makePathName}
\alias{makePathName}
\title{create appropriate directory structure if needed}
\description{create appropriate directory structure if needed}
\usage{makePathName(path, MakePath = TRUE, verbose = 0)}
\arguments{
  \item{path}{path to create}
  \item{MakePath}{if yes, create directory if not exists}
  \item{verbose}{level of verbosity}
}

\value{returns absolute path}

\author{"Markus Loecher, Berlin School of Economics and Law (BSEL)"  <markus.loecher@gmail.com>}





