\name{MyReportBegin}
\alias{MyReportBegin}
\title{gracefully initializes the HTML page}
\description{gracefully initializes the HTML page}
\usage{MyReportBegin(file = "report.html", title = "My Report Title", 
    header = NULL)}
\arguments{
  \item{file}{filename}
  \item{title}{title for HTML page}
  \item{header}{header yes/no}
}



\author{"Markus Loecher, Berlin School of Economics and Law (BSEL)"  <markus.loecher@gmail.com>}





