\name{MyReportEnd}
\alias{MyReportEnd}
\title{gracefully finalizes the HTML page}
\description{gracefully finalizes the HTML page}
\usage{MyReportEnd(file = "report.html")}
\arguments{
  \item{file}{file to append to}
}



\author{"Markus Loecher, Berlin School of Economics and Law (BSEL)" <markus.loecher@gmail.com>}





